[consul] Add data dependencies to profile
Allow resources to fail if they have insufficient data.
This is preferable to catalog fail() functions because
it allows the rest of the catalog to apply as normal.

Without this feauture, a node providing services that
depend on each other will never complete since it will
fail to compile.

Uses the orhcestration_utils module.