[consul] Allow data to determine inclusions
Sometimes failing a class using runtime_fail will cause a lot of
nasty side effects where many profiles are on a single host and
there are cross profile orderings. An example of this is where
nova::keystone::auth needs to know nova-api_Address, but is also
set before nova-api service in the dependency chain. This means we
can't set up the service before consuming its output, which is
extremely problematic.

This patch adds the ability to control when classes are included,
so that if nova-api_Address is not available in the above example,
nova::keystone::auth will not be included in the catalog